implement player relocation functionality
